test: cover state normalization
